chore: rename `custom.events` to `custom.autocmds`
